+++ rmi-src/rmi/ServeurRMI.java
@@ -5,6 +5,6 @@ import java.rmi.RemoteException;
 
 public interface ServeurRMI extends Remote {
 	public String meth() throws RemoteException;
-	public void newImage(String title, Integer size) throws RemoteException;
-	public void insertByte(String title, Integer begin, byte[] buffer) throws RemoteException;
+	public void newImage(String title, Long size) throws RemoteException;
+	public void insertByte(String title, Long begin, byte[] buffer) throws RemoteException;
 }
